script/requirements.txt
pypinyin==0.47.1
Pillow==9.4.0